+++ ase/constraints.py
@@ -4,6 +4,7 @@ from warnings import warn
 
 import numpy as np
 
+from ase import Atoms
 from ase.filters import ExpCellFilter as ExpCellFilterOld
 from ase.filters import Filter as FilterOld
 from ase.filters import StrainFilter as StrainFilterOld
@@ -55,7 +56,7 @@ def constrained_indices(atoms, only_include=None):
 class FixConstraint:
     """Base class for classes that fix one or more atoms in some way."""
 
-    def index_shuffle(self, atoms, ind):
+    def index_shuffle(self, atoms: Atoms, ind):
         """Change the indices.
 
         When the ordering of the atoms in the Atoms object changes,
@@ -67,7 +68,7 @@ class FixConstraint:
         """
         raise NotImplementedError
 
-    def repeat(self, m, n):
+    def repeat(self, m: int, n: int):
         """ basic method to multiply by m, needs to know the length
         of the underlying atoms object for the assignment of
         multiplied constraints to work.
@@ -77,12 +78,27 @@ class FixConstraint:
                'remove your constraints.')
         raise NotImplementedError(msg)
 
-    def adjust_momenta(self, atoms, momenta):
-        """Adjusts momenta in identical manner to forces."""
+    def get_removed_dof(self, atoms: Atoms):
+        """Get number of removed degrees of freedom due to constraint."""
+
+    def adjust_positions(self, atoms: Atoms, new):
+        """Adjust positions."""
+
+    def adjust_momenta(self, atoms: Atoms, momenta):
+        """Adjust momenta."""
+        # The default is in identical manner to forces.
+        # TODO: The default is however not always reasonable.
         self.adjust_forces(atoms, momenta)
 
+    def adjust_forces(self, atoms: Atoms, forces):
+        """Adjust forces."""
+
     def copy(self):
+        """Copy constraint."""
         return dict2constraint(self.todict().copy())
+
+    def todict(self):
+        """Convert constraint to dictionary."""
 
 
 class IndexedConstraint(FixConstraint):
@@ -789,53 +805,71 @@ class FixedLine(IndexedConstraint):
 
 
 class FixCartesian(IndexedConstraint):
-    'Fix an atom index *a* in the directions of the cartesian coordinates.'
+    """Fix atoms in the directions of the cartesian coordinates.
 
-    def __init__(self, a, mask=(1, 1, 1)):
+    Parameters
+    ----------
+    a : Sequence[int]
+        Indices of atoms to be fixed.
+    mask : tuple[bool, bool, bool], default: (True, True, True)
+        Cartesian directions to be fixed. (False: unfixed, True: fixed)
+    """
+
+    def __init__(self, a, mask=(True, True, True)):
         super().__init__(indices=a)
-        self.mask = ~np.asarray(mask, bool)
+        self.mask = np.asarray(mask, bool)
 
-    def get_removed_dof(self, atoms):
-        return (3 - self.mask.sum()) * len(self.index)
+    def get_removed_dof(self, atoms: Atoms):
+        return self.mask.sum() * len(self.index)
 
-    def adjust_positions(self, atoms, new):
-        step = new[self.index] - atoms.positions[self.index]
-        step *= self.mask[None, :]
-        new[self.index] = atoms.positions[self.index] + step
+    def adjust_positions(self, atoms: Atoms, new):
+        new[self.index] = np.where(
+            self.mask[None, :],
+            atoms.positions[self.index],
+            new[self.index],
+        )
 
-    def adjust_forces(self, atoms, forces):
-        forces[self.index] *= self.mask[None, :]
-
-    def __repr__(self):
-        return 'FixCartesian(indices={}, mask={})'.format(
-            self.index.tolist(), list(~self.mask))
+    def adjust_forces(self, atoms: Atoms, forces):
+        forces[self.index] *= ~self.mask[None, :]
 
     def todict(self):
         return {'name': 'FixCartesian',
                 'kwargs': {'a': self.index.tolist(),
-                           'mask': (~self.mask).tolist()}}
+                           'mask': self.mask.tolist()}}
+
+    def __repr__(self):
+        name = type(self).__name__
+        return f'{name}(indices={self.index.tolist()}, {self.mask.tolist()})'
 
 
 class FixScaled(IndexedConstraint):
-    'Fix an atom index *a* in the directions of the unit vectors.'
+    """Fix atoms in the directions of the unit vectors.
 
-    def __init__(self, a, mask=(1, 1, 1), cell=None):
+    Parameters
+    ----------
+    a : Sequence[int]
+        Indices of atoms to be fixed.
+    mask : tuple[bool, bool, bool], default: (True, True, True)
+        Cell directions to be fixed. (False: unfixed, True: fixed)
+    """
+
+    def __init__(self, a, mask=(True, True, True), cell=None):
         # XXX The unused cell keyword is there for compatibility
         # with old trajectory files.
-        super().__init__(a)
-        self.mask = np.array(mask, bool)
+        super().__init__(indices=a)
+        self.mask = np.asarray(mask, bool)
 
-    def get_removed_dof(self, atoms):
+    def get_removed_dof(self, atoms: Atoms):
         return self.mask.sum() * len(self.index)
 
-    def adjust_positions(self, atoms, new):
+    def adjust_positions(self, atoms: Atoms, new):
         cell = atoms.cell
         scaled_old = cell.scaled_positions(atoms.positions[self.index])
         scaled_new = cell.scaled_positions(new[self.index])
         scaled_new[:, self.mask] = scaled_old[:, self.mask]
         new[self.index] = cell.cartesian_positions(scaled_new)
 
-    def adjust_forces(self, atoms, forces):
+    def adjust_forces(self, atoms: Atoms, forces):
         # Forces are covariant to the coordinate transformation,
         # use the inverse transformations
         cell = atoms.cell
@@ -849,7 +883,8 @@ class FixScaled(IndexedConstraint):
                            'mask': self.mask.tolist()}}
 
     def __repr__(self):
-        return f'FixScaled({self.index.tolist()}, {self.mask})'
+        name = type(self).__name__
+        return f'{name}(indices={self.index.tolist()}, {self.mask.tolist()})'
 
 
 # TODO: Better interface might be to use dictionaries in place of very